add `if` modifier to be used on `.validates` calls
allows `validates :foo, ... , if: -> (record) { record.some_criteria? }`

meaning validation can be run or not run depending on state of object being validated.

This is added at the top level, so option is available to all validators.

If multiple validators are specified in a single `validates` call, the `if` modfiier applies to all of them.